refactor: move RelayModel interface to types/
