+++ types/relay-model.ts
@@ -1,0 +1,9 @@
+export interface RelayModel {
+	modelKey: string;
+	name: string;
+	tagline: string;
+	genres: string;
+	description: string;
+	status: 'available' | 'planned';
+	href?: string;
+}
